# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grainily API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List published ads
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: product
          in: query
          schema:
            type: string
        - name: quantity
          in: query
          schema:
            type: string
        - name: unit
          in: query
          schema:
            type: string
        - name: listingType
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, oldest, price_asc, price_desc]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Submit a new ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Created ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}:
    get:
      operationId: getAd
      tags: [ads]
      summary: Get ad by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
...
  /admin/ads:
    get:
      operationId: adminListAds
      tags: [admin]
      summary: Admin - list all ads (including pending)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
  /admin/ads/{id}/status:
    patch:
      operationId: updateAdStatus
      tags: [admin]
      summary: Admin - update ad status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdStatusUpdate"
      responses:
        "200":
          description: Updated ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /admin/ads/{id}:
    patch:
      operationId: updateAd
      tags: [admin]
      summary: Admin - edit any field of an ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdUpdate"
      responses:
        "200":
          description: Updated ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    delete:
      operationId: deleteAd
      tags: [admin]
      summary: Admin - delete ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/ads/bulk:
    post:
      operationId: adminBulkUpdateAds
      tags: [admin]
      summary: Admin - bulk approve / reject / delete ads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAdAction"
      responses:
        "200":
          description: Number of ads affected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResult"
...
  /admin/advertisements:
    get:
      operationId: adminListAdvertisements
      tags: [admin]
      summary: Admin - list all advertisements
      responses:
        "200":
          description: All advertisements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Advertisement"
    post:
      operationId: adminCreateAdvertisement
      tags: [admin]
      summary: Admin - create advertisement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementInput"
      responses:
        "201":
          description: Created advertisement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Advertisement"
  /admin/advertisements/reorder:
    post:
      operationId: adminReorderAdvertisements
      tags: [admin]
      summary: Admin - reorder advertisements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementReorderInput"
      responses:
        "200":
          description: Reordered advertisements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Advertisement"
  /admin/advertisements/{id}:
    put:
      operationId: adminUpdateAdvertisement
      tags: [admin]
      summary: Admin - update advertisement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertisementInput"
      responses:
        "200":
          description: Updated advertisement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Advertisement"
    delete:
      operationId: adminDeleteAdvertisement
      tags: [admin]
      summary: Admin - delete advertisement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Ad:
      type: object
      required: [id, title, location, product, category, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        product:
          type: string
        quantity:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        category:
          type: string
        listingType:
          type: string
          enum: [free, flexible, fixed]
        price:
          type: ["string", "null"]
        isPromoted:
          type: boolean
        promotionDuration:
          type: ["integer", "null"]
        promotionPrice:
          type: ["string", "null"]
        subscriptionType:
          type: string
          enum: [none, weekly, monthly, annual]
        subscriptionPrice:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, published, rejected]
        createdAt:
          type: string
        lastEditedByAdmin:
          type: ["string", "null"]
        lastEditedAt:
          type: ["string", "null"]
    AdInput:
      type: object
      required: [title, location, product, category]
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        product:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        listingType:
          type: string
          enum: [free, flexible, fixed]
        price:
          type: string
        isPromoted:
          type: boolean
        promotionDuration:
          type: integer
        promotionPrice:
          type: string
        subscriptionType:
          type: string
          enum: [none, weekly, monthly, annual]
        subscriptionPrice:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
    AdStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, published, rejected]
    AdUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        product:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        listingType:
          type: string
          enum: [free, flexible, fixed]
        price:
          type: string
        isPromoted:
          type: boolean
        promotionDuration:
          type: integer
        promotionPrice:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
        status:
          type: string
          enum: [pending, published, rejected]
    Advertisement:
      type: object
      required: [id, title, mediaType, mediaUrl, displayOrder, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        mediaUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        displayOrder:
          type: integer
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
    AdvertisementInput:
      type: object
      required: [title, mediaType, mediaUrl]
      properties:
        title:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        mediaUrl:
          type: string
        linkUrl:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    AdvertisementReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
...
    BulkAdAction:
      type: object
      required: [ids, action]
      properties:
        ids:
          type: array
          items:
            type: integer
        action:
          type: string
          enum: [publish, reject, delete]
    BulkResult:
      type: object
      required: [affected]
      properties:
        affected:
          type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT